feat: 커뮤니티 api 연동
