+++ src/features/community/components/CommentInput.tsx
@@ -1,11 +1,12 @@
 'use client';
 
-import { useEffect, useState } from 'react';
+import { useEffect, useId, useState } from 'react';
 
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import * as Switch from '@radix-ui/react-switch';
 import { Controller, useForm } from 'react-hook-form';
+
+import { Check } from '@common/components/svg/Check';
 
 import queryKeys from '@common/constants/queryKeys';
 
@@ -15,6 +16,7 @@ import { type CreateCommunityCommentRequestDTO } from '../schema';
 interface CommentInputProps {
   articleId: string;
   parentCommentId?: string | null;
+  onReplyCancel?: () => void;
 }
 
 type CommentFormValues = Pick<CreateCommunityCommentRequestDTO, 'content' | 'isAnonymous'>;
@@ -24,9 +26,14 @@ const DEFAULT_VALUES: CommentFormValues = {
   isAnonymous: true,
 };
 
-export default function CommentInput({ articleId, parentCommentId = null }: CommentInputProps) {
+export default function CommentInput({
+  articleId,
+  parentCommentId = null,
+  onReplyCancel,
+}: CommentInputProps) {
   const queryClient = useQueryClient();
   const [submitError, setSubmitError] = useState<string | null>(null);
+  const checkboxId = useId();
   const {
     register,
     control,
@@ -51,6 +58,7 @@ export default function CommentInput({ articleId, parentCommentId = null }: Comm
       });
       reset(DEFAULT_VALUES);
       setSubmitError(null);
+      onReplyCancel?.();
     },
     onError: (error) => {
       const message = error instanceof Error ? error.message : '댓글 작성 중 오류가 발생했습니다.';
@@ -68,35 +76,45 @@ export default function CommentInput({ articleId, parentCommentId = null }: Comm
   });
 
   return (
-    <form
+    <div className='sticky bottom-0 left-0 right-0'>
+      <form
       onSubmit={onSubmit}
-      className="flex w-full flex-col gap-4 rounded-2xl border border-gray-100 bg-white p-6 shadow-sm"
+      className="flex w-full flex-col bg-gray-50 px-[14px] py-[12px] rounded-[16px]"
     >
-      <div className="flex items-center justify-between">
-        <span className="text-p15b text-gray-900">익명</span>
+      <div className="flex items-start justify-between gap-[12px]">
         <Controller
           control={control}
           name="isAnonymous"
           render={({ field: { value, onChange } }) => (
-            <label className="flex items-center gap-2">
-              <span className="text-p13 text-gray-500">익명으로 작성</span>
-              <Switch.Root
-                className="relative h-5 w-9 rounded-full bg-gray-200 data-[state=checked]:bg-gray-900"
-                checked={value}
-                onCheckedChange={onChange}
-              >
-                <Switch.Thumb className="block h-4 w-4 translate-x-1 rounded-full bg-white transition-transform data-[state=checked]:translate-x-[18px]" />
-              </Switch.Root>
+            <label htmlFor={checkboxId} className="flex items-center gap-[5px] cursor-pointer">
+              <div className="relative">
+                <input
+                  id={checkboxId}
+                  type="checkbox"
+                  className="absolute inset-0 cursor-pointer opacity-0"
+                  checked={!!value}
+                  onChange={(event) => onChange(event.target.checked)}
+                />
+                <div
+                  className={`size-[16px] rounded flex items-center justify-center transition-colors duration-200 ${
+                    value
+                      ? 'border-gray-900 bg-gray-900'
+                      : 'border-gray-200 bg-gray-600'
+                  }`}
+                >
+                  {value && <Check className="size-4 text-white" />}
+                </div>
+              </div>
+              <span className="text-p15b text-gray-700 whitespace-nowrap">익명</span>
             </label>
           )}
         />
+        <textarea
+          className="h-46px w-full resize-none text-gray-900 placeholder:text-gray-200 focus:outline-none"
+          placeholder="따뜻한 댓글을 입력해주세요 :)"
+          {...register('content', { required: '댓글을 입력해주세요.' })}
+        />
       </div>
-      <textarea
-        rows={3}
-        className="text-p15 w-full resize-none rounded-xl border border-gray-200 p-4 text-gray-700 outline-none focus:border-gray-900"
-        placeholder="따뜻한 댓글을 입력해주세요 :)"
-        {...register('content', { required: '댓글을 입력해주세요.' })}
-      />
       {errors.content && (
         <p className="text-p13 text-red-500">{errors.content.message ?? '댓글을 입력해주세요.'}</p>
       )}
@@ -105,11 +123,12 @@ export default function CommentInput({ articleId, parentCommentId = null }: Comm
         <button
           type="submit"
           disabled={isPending}
-          className="text-p15b flex items-center rounded-lg bg-gray-900 px-5 py-2 text-white transition-opacity disabled:opacity-50"
+          className="text-p16b px-[12px] py-[4px] text-gray-700 transition-opacity disabled:opacity-50"
         >
           {isPending ? '등록 중...' : '등록'}
         </button>
       </div>
     </form>
+    </div>
   );
 }
